+++ osc-server/src/main/java/org/osc/core/broker/model/plugin/ApiFactoryService.java
@@ -49,8 +49,6 @@ import org.osgi.annotation.versioning.ConsumerType;
 
 @ConsumerType
 public interface ApiFactoryService {
-    String MANAGER_PLUGINS_DIRECTORY = "mgr_plugins";
-    String SDN_CONTROLLER_PLUGINS_DIRECTORY = "sdn_ctrl_plugins";
     String PLUGINS_DIRECTORY = "plugins";
 
     ApplianceManagerApi createApplianceManagerApi(ManagerType managerType) throws Exception;
