--- osc-service-api/src/main/java/org/osc/core/broker/service/api/ImportSdnControllerPluginServiceApi.java
@@ -1,26 +1,0 @@
-/*******************************************************************************
- * Copyright (c) Intel Corporation
- * Copyright (c) 2017
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *******************************************************************************/
-package org.osc.core.broker.service.api;
-
-import org.osc.core.broker.service.request.ImportFileRequest;
-import org.osc.core.broker.service.response.BaseResponse;
-
-public interface ImportSdnControllerPluginServiceApi
-        extends ServiceDispatcherApi<ImportFileRequest, BaseResponse> {
-
-    boolean isControllerTypeUsed(String name);
-}
