+++ osc-ui/src/main/java/org/osc/core/broker/view/maintenance/PluginsLayout.java
@@ -53,7 +53,7 @@ import com.vaadin.ui.Table;
 import com.vaadin.ui.UI;
 import com.vaadin.ui.VerticalLayout;
 
-public class ManagerPluginsLayout extends FormLayout {
+public class PluginsLayout extends FormLayout {
 
     private static final String PROP_PLUGIN_INFO = "Info";
     private static final String PROP_PLUGIN_STATE = "State";
@@ -64,7 +64,7 @@ public class ManagerPluginsLayout extends FormLayout {
 
     private static final long serialVersionUID = 1L;
 
-    Logger log = Logger.getLogger(ManagerPluginsLayout.class);
+    Logger log = Logger.getLogger(PluginsLayout.class);
 
     private Table plugins;
     private Panel pluginsPanel;
@@ -73,14 +73,12 @@ public class ManagerPluginsLayout extends FormLayout {
     private ImportPluginServiceApi importPluginService;
     private ServiceRegistration<PluginListener> registration;
 
-    public ManagerPluginsLayout(BundleContext ctx,
-    		ImportPluginServiceApi importPluginService,
+    public PluginsLayout(BundleContext ctx, ImportPluginServiceApi importPluginService,
             ServerApi server) throws Exception {
         super();
         this.importPluginService = importPluginService;
-        this.uploader = new PluginUploader(PluginType.MANAGER, server);
 
-        // Create Controls
+        this.uploader = new PluginUploader(PluginType.SDN, server);
         VerticalLayout uploadContainer = new VerticalLayout();
         VerticalLayout pluginsContainer = new VerticalLayout();
         VerticalLayout sdkContainer = new VerticalLayout();
@@ -118,49 +116,52 @@ public class ManagerPluginsLayout extends FormLayout {
         addComponent(pluginsContainer);
         addComponent(sdkContainer);
 
+
         // Subscribe to Plugin Notifications
         this.registration = ctx.registerService(PluginListener.class,
-            this::updateTable, null);
+                this::updateTable, null);
     }
 
-    private void updateTable(PluginEvent ev) {
-            PluginApi plugin = ev.getPlugin();
-            if(plugin.getType() != PluginType.MANAGER) {
-                return;
-            }
+    private void updateTable(PluginEvent event) {
+        PluginApi plugin = event.getPlugin();
 
-            switch (ev.getType()) {
-            case ADDING:
-                Item addingItem = this.plugins.addItem(plugin);
-                if (addingItem != null) {
-                    updateItem(addingItem, plugin);
-                }
-                break;
-            case MODIFIED:
-                Item modifyingItem = this.plugins.getItem(plugin);
-                if (modifyingItem == null) {
-                    modifyingItem = this.plugins.addItem(plugin);
-                }
-                if (modifyingItem != null) {
-                    updateItem(modifyingItem, plugin);
-                }
-                break;
-            case REMOVED:
-                this.plugins.removeItem(plugin);
-                break;
-            default:
-            	this.log.error("Unknown plugin event type: " + ev.getType());
-            	break;
+        PluginType pluginType = plugin.getType();
+        
+        if(! ((PluginType.SDN.equals(pluginType))  || (PluginType.MANAGER.equals(pluginType)))) {
+        	return;
+        }
+        
+        switch (event.getType()) {
+        case ADDING:
+            Item addingItem = this.plugins.addItem(plugin);
+            if (addingItem != null) {
+                updateItem(addingItem, plugin);
             }
+            break;
+        case MODIFIED:
+            Item modifyingItem = this.plugins.getItem(plugin);
+            if (modifyingItem == null) {
+                modifyingItem = this.plugins.addItem(plugin);
+            }
+            if (modifyingItem != null) {
+                updateItem(modifyingItem, plugin);
+            }
+            break;
+        case REMOVED:
+            this.plugins.removeItem(plugin);
+            break;
+        default:
+        	this.log.error("Unknown plugin event type: " + event.getType());
+        	break;
+        }
     }
 
     private Button getDownloadSdkButton() throws URISyntaxException, MalformedURLException {
         SdkUtil sdkUtil = new SdkUtil();
-        Button downloadSdk = new Button(VmidcMessages.getString(VmidcMessages_.MAINTENANCE_MANAGERPLUGIN_DOWNLOAD_SDK));
+        Button downloadSdk = new Button(VmidcMessages.getString(VmidcMessages_.MAINTENANCE_SDNPLUGIN_DOWNLOAD_SDK));
         URI currentLocation = UI.getCurrent().getPage().getLocation();
-
         URI downloadLocation = new URI(currentLocation.getScheme(), null, currentLocation.getHost(),
-                currentLocation.getPort(), sdkUtil.getSdk(SdkUtil.sdkType.MANAGER), null, null);
+                currentLocation.getPort(), sdkUtil.getSdk(SdkUtil.sdkType.SDN_CONTROLLER), null, null);
         FileDownloader downloader = new FileDownloader(new ExternalResource(downloadLocation.toURL().toString()));
         downloader.extend(downloadSdk);
         return downloadSdk;
@@ -190,8 +191,8 @@ public class ManagerPluginsLayout extends FormLayout {
     private void deletePlugin(PluginApi plugin) {
         final VmidcWindow<OkCancelButtonModel> deleteWindow = WindowUtil.createAlertWindow("Delete Plugin", "Delete Plugin - " + plugin.getSymbolicName());
         deleteWindow.getComponentModel().setOkClickedListener(event -> {
-            if (this.importPluginService.isManagerTypeUsed(plugin.getName())) {
-                ViewUtil.iscNotification("Manager Plugin '" + plugin.getName() + "' is used.", Notification.Type.ERROR_MESSAGE);
+            if (this.importPluginService.isControllerTypeUsed(plugin.getName())) {
+                ViewUtil.iscNotification("SDN Controller Plugin '" + plugin.getName() + "' is used.", Notification.Type.ERROR_MESSAGE);
             } else {
                 try {
                     File origin = plugin.getOrigin();
@@ -202,7 +203,7 @@ public class ManagerPluginsLayout extends FormLayout {
                     // Use Java 7 Files.delete(), as it throws an informative exception when deletion fails
                     Files.delete(origin.toPath());
                 } catch (Exception e) {
-                    ViewUtil.showError("Fail to remove Manager Plugin '" + plugin.getSymbolicName() + "'", e);
+                    ViewUtil.showError("Fail to remove Plugin '" + plugin.getSymbolicName() + "'", e);
                 }
             }
             deleteWindow.close();
@@ -218,13 +219,13 @@ public class ManagerPluginsLayout extends FormLayout {
                 try {
                     ImportFileRequest importRequest = new ImportFileRequest(uploadPath);
 
-                    ManagerPluginsLayout.this.importPluginService.dispatch(importRequest);
+                    PluginsLayout.this.importPluginService.dispatch(importRequest);
 
-                    ViewUtil.iscNotification(VmidcMessages.getString(VmidcMessages_.UPLOAD_PLUGIN_MANAGER_SUCCESSFUL),
-                            null, Notification.Type.TRAY_NOTIFICATION);
-
+                    ViewUtil.iscNotification(
+                            VmidcMessages.getString(VmidcMessages_.UPLOAD_PLUGIN_SDNCONTROLLER_SUCCESSFUL), null,
+                            Notification.Type.TRAY_NOTIFICATION);
                 } catch (Exception e) {
-                    ManagerPluginsLayout.this.log.info(e.getMessage());
+                    PluginsLayout.this.log.info(e.getMessage());
                     ViewUtil.iscNotification(e.getMessage(), Notification.Type.ERROR_MESSAGE);
                 }
             }
